refactor(@movemate/backend): files restructured
